+++ tests/testthat/test-name_repair.R
@@ -141,6 +141,41 @@ test_that("unique-ification has an 'algebraic'-y property", {
   expect_identical(z1, z4)
 })
 
+# names<-() -------------------------------------------------------------------
+test_that("names<-() and set_names() reject non-minimal names", {
+  df <- tibble(a = 1:3, b = 4:6, c = 7:9)
+
+  expect_error(
+    set_names(df, NULL),
+    error_names_must_be_non_null(),
+    fixed = TRUE
+  )
+  expect_error(
+    `names<-`(df, NA),
+    error_names_must_have_length(1, 3),
+    fixed = TRUE
+  )
+  expect_error(
+    `names<-`(df, ""),
+    error_names_must_have_length(1, 3),
+    fixed = TRUE
+  )
+  expect_error(
+    `names<-`(df, 1),
+    error_names_must_have_length(1, 3),
+    fixed = TRUE
+  )
+  expect_error(
+    `names<-`(df, 1:2),
+    error_names_must_have_length(2, 3),
+    fixed = TRUE
+  )
+  expect_identical(
+    set_names(df, letters[1:3]),
+    tibble(a = 1:3, b = 4:6, c = 7:9)
+  )
+})
+
 # repair_names (deprecated) ---------------------------------------------------
 test_that("zero-length inputs given character names", {
   out <- repair_names(character())