+++ NEWS.md
@@ -10,6 +10,8 @@ The `tibble()` and `as_tibble()` functions, and the low-level `new_tibble()` con
 
 - The `nrow` argument to `new_tibble()` is now mandatory.
 
+- Setting names on a tibble via `names(df) <- ...` now also requires minimal names (#466).
+
 - In `as_tibble()`, checking names is also enabled by default, even for tibbles, matrices and other matrix-like objects: names must exist, `NA` names are not allowed. In particular, coercing a matrix without column names will trigger an error. (This corresponds to the `"minimal"` checks described below.). The `validate` argument to `as_tibble()` has been deprecated. (The `as_tibble.tbl_df()` method has been removed, the `as_tibble.data.frame()` method will be used for tibbles.)
 
 - `as_tibble()` always checks that all columns are 1D or 2D vectors and not of type `POSIXlt`, even with `validate = FALSE` (which is now deprecated).